refactor ContactManager class
